+++ CHANGELOG.md
@@ -7,6 +7,10 @@ All notable changes to this project will be documented in this file. This projec
 ### Added
 
 - (input) support for **ng-pattern** as **templateOptions.pattern** (#10)
+- (chips) has now **label** wrapper
+- (datepicker) has now **label** wrapper
+- (radio) has now **label** wrapper
+- (slider) has now **label** wrapper
 
 ## [0.12.0](https://github.com/formly-js/angular-formly-templates-material/compare/v0.11.0...v0.12.0) - 2015-12-30